# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amazon Experts Backend API",
    "description": "Swagger/OpenAPI documentation for the Amazon Experts backend service. Platform APIs do not require x-tenant-id; tenant APIs require x-tenant-id in the request header.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/experts": {
      "get": {
        "tags": [
          "experts"
        ],
        "summary": "List experts",
        "operationId": "list_experts_api_v1_experts_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoryid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "published",
                    "draft",
                    "unlisted"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "experts"
        ],
        "summary": "Create an expert",
        "operationId": "create_expert_api_v1_experts_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpertRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expert"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/experts/{expert_id}": {
      "get": {
        "tags": [
          "experts"
        ],
        "summary": "Get expert details",
        "operationId": "get_expert_api_v1_experts__expert_id__get",
        "parameters": [
          {
            "name": "expert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Expert Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expert"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "experts"
        ],
        "summary": "Update an expert",
        "operationId": "update_expert_api_v1_experts__expert_id__patch",
        "parameters": [
          {
            "name": "expert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Expert Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expert"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "experts"
        ],
        "summary": "Delete an expert",
        "operationId": "delete_expert_api_v1_experts__expert_id__delete",
        "parameters": [
          {
            "name": "expert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Expert Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/expert-market/categories": {
      "get": {
        "tags": [
          "expert-market"
        ],
        "summary": "List public expert categories",
        "operationId": "list_market_categories_api_v1_expert_market_categories_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertMarketCategoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/expert-market/experts": {
      "get": {
        "tags": [
          "expert-market"
        ],
        "summary": "List public experts",
        "operationId": "list_market_experts_api_v1_expert_market_experts_get",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoryid"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertMarketExpertListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/expert-market/experts/{expert_id}": {
      "get": {
        "tags": [
          "expert-market"
        ],
        "summary": "Get public expert details",
        "operationId": "get_market_expert_api_v1_expert_market_experts__expert_id__get",
        "parameters": [
          {
            "name": "expert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Expert Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertMarketExpert"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateExpertRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categoryId": {
            "type": "string",
            "title": "Categoryid"
          },
          "abilityIntro": {
            "type": "string",
            "title": "Abilityintro"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "status": {
            "type": "string",
            "enum": [
              "published",
              "draft",
              "unlisted"
            ],
            "title": "Status",
            "default": "draft"
          },
          "skillIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skillids"
          },
          "knowledgeBaseIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledgebaseids"
          },
          "guideQuestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Guidequestions"
          },
          "summonButtonText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summonbuttontext"
          }
        },
        "type": "object",
        "required": [
          "name",
          "categoryId",
          "abilityIntro"
        ],
        "title": "CreateExpertRequest"
      },
...
      "Expert": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categoryId": {
            "type": "string",
            "title": "Categoryid"
          },
          "categoryName": {
            "type": "string",
            "title": "Categoryname"
          },
          "abilityIntro": {
            "type": "string",
            "title": "Abilityintro"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "status": {
            "type": "string",
            "enum": [
              "published",
              "draft",
              "unlisted"
            ],
            "title": "Status"
          },
          "skillIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skillids"
          },
          "knowledgeBaseIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledgebaseids"
          },
          "guideQuestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Guidequestions"
          },
          "summonButtonText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summonbuttontext"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "categoryId",
          "categoryName",
          "abilityIntro",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "Expert"
      },
...
      "ExpertListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Expert"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ExpertListResponse"
      },
      "ExpertMarketCategory": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ExpertMarketCategory"
      },
      "ExpertMarketCategoryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExpertMarketCategory"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ExpertMarketCategoryListResponse"
      },
      "ExpertMarketExpert": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categoryId": {
            "type": "string",
            "title": "Categoryid"
          },
          "categoryName": {
            "type": "string",
            "title": "Categoryname"
          },
          "abilityIntro": {
            "type": "string",
            "title": "Abilityintro"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "guideQuestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Guidequestions"
          },
          "summonButtonText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summonbuttontext"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "categoryId",
          "categoryName",
          "abilityIntro"
        ],
        "title": "ExpertMarketExpert"
      },
      "ExpertMarketExpertListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExpertMarketExpert"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ExpertMarketExpertListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateExpertRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "categoryId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoryid"
          },
          "abilityIntro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abilityintro"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "skillIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skillids"
          },
          "knowledgeBaseIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledgebaseids"
          },
          "guideQuestions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Guidequestions"
          },
          "summonButtonText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summonbuttontext"
          }
        },
        "type": "object",
        "title": "UpdateExpertRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}